Add more commands, support for player id lookup, and support 0xFF selection_count
Adds GameContext object to keep track of state used for parsing. Namely:
- Keep track of which player owns any object IDs we see, so we can use this
information to assign a player_id to commands that have only object IDs
- Keep track of the previous selected_ids, so that we can fill it in when
it is repeated (a selection_count of 0xFF means the selection is the same
as the previous command). Note, there's some concern here because I don't
know if there is any other scope applied. In a multiplayer context, it makes
sense for each client to use this rule, so it could be that what is stored
in the log actually means "use the selected_ids in the last command sent by
this player".